CRITICAL SEO: Convert homepage and how-owner-finance-works to server components with full SEO optimization, schema markup, and proper metadata
